Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/error-type-3-expected.txt

CONSOLE MESSAGE: line 2: [object Object]
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.

FAIL instantiation error has higher priority than evaluation error null is not an object (evaluating 'log[3].constructor')
